requirements.txt
bcrypt==4.0.1
blinker==1.6.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.1.0
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
fonttools==4.40.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
numpy==1.25.0
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==10.0.0
preshed==3.0.8
pydantic==1.10.11
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.16
srsly==2.4.6
sympy==1.12
thinc==8.1.10
threadpoolctl==3.1.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wasabi==1.1.2
Werkzeug==2.3.6
zipp==3.15.0
